simplify the screen buffer features.
for now just use a static buffer size of 20x4
and disable the windowing abilities.

it would seem dynamically allocating the memory and using string copy functions was causing
us issues where memory would be corrupted, or reading from the buffer would fetch more than
it should. Remember the LCD screen is not a simple null terminated string. It is just an
array of characters.